pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job_hunter"
version = "0.1.0"
description = "Job Hunter: fetcher annunci + DB + email"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.2"
]

[project.scripts]
job-hunter = "main:cli"
job-hunter-batch = "run_batch:main"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["main", "fetch_remotive", "db", "report", "notify", "run_batch"]
